refactoring

+++ src/utils/getinttype.jl
@@ -1,7 +1,7 @@
-function get_int_type(n)
+function get_int_type(n::Integer)
     return eval(Symbol("Int", n))
 end
 
-function get_uint_type(n)
+function get_uint_type(n::Integer)
     return eval(Symbol("UInt", n))
 end
